requirements.txt
Flask==1.1.2
flask-accept==0.0.6
Flask-API==2.0
Flask-Caching==1.8.0
Flask-Cors==3.0.8
requests==2.23.0
python-dotenv==0.14.0